requirements.txt
mne
tqdm
h5py
xlrd
numpy
scipy
pandas
scikit-learn
xgboost
seaborn
jupyter
openpyxl
kendall-w
quantities
matplotlib
ruamel.yaml
hrv-analysis
tensorflow-addons
jupyter_http_over_ws
